cambio contraproducente en la estructura

+++ acad-apoderado-service/src/main/java/com/acad/apoderadoservice/Entity/Apoderado.java
@@ -17,7 +17,7 @@ public class Apoderado {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "idApoderado")
-    private Long idApoderado;
+    private Integer idApoderado;
 
     @Column(name = "nombres", length = 45)
     private String nombres;
